requirements/dev.txt
-r base.txt
django-debug-toolbar>=4.3
factory-boy>=3.3
pytest-django>=4.8
coverage>=7.4
django-extensions>=3.2
